added timer and player name data
A timer was added, it starts counting down from 6:00 to 0:00 and stops (no functionality after 0:00). I also made it to where the names of the players on each team is returned from player_screen and made as an argument in action_screen. Then I added a dictionary for scores and set them to 0.